Applying Brian’s appstore callback fix
https://github.com/bgeihsgt/react-native-linkedin-login/commit/582b6aaba7baf68ebaa7993a2b16054891f19843

Fix android issue where the app store prompt never calls back.
If the LinkedIn SDK detects that the LinkedIn app is not installed, it
never calls the auth callback supplied to the function. Thus clients are
just left hanging.

This will fail the callback with an app not installed error.